feat: splitted stream patcher
