--- src/utils/stream-patcher.ts
@@ -1,1149 +1,0 @@
-import {
-  StreamCropConfig,
-  StreamFilterConfig,
-  StreamPatcherSettings,
-  StreamPatcherSize,
-} from "@/types/stream-patcher";
-import { CanvasManager, type CanvasResource } from "./canvas-manager";
-import { ConfettiOverlayManager } from "./confetti-overlay";
-import { Logger } from "./log";
-import { MediaOverlayManager } from "./media-overlay";
-import { GlobalPerformanceMonitor } from "./performance-monitor";
-
-function normalizeFilterValue(value?: number): number {
-  if (typeof value !== "number") return 1;
-  return Math.max(0, value) / 100;
-}
-
-/**
- * Performance Monitor for tracking rendering metrics
- */
-class PerformanceMonitor {
-  private frameCount = 0;
-  private lastFpsTime = performance.now();
-  private currentFps = 0;
-  private dropCount = 0;
-  private renderTimes: number[] = [];
-
-  startFrame(): number {
-    return performance.now();
-  }
-
-  endFrame(startTime: number): void {
-    const renderTime = performance.now() - startTime;
-    this.renderTimes.push(renderTime);
-
-    // Keep only last 60 samples
-    if (this.renderTimes.length > 60) {
-      this.renderTimes.shift();
-    }
-
-    this.frameCount++;
-    const now = performance.now();
-
-    if (now - this.lastFpsTime >= 1000) {
-      this.currentFps = this.frameCount;
-      this.frameCount = 0;
-      this.lastFpsTime = now;
-    }
-  }
-
-  recordDrop(): void {
-    this.dropCount++;
-  }
-
-  getStats() {
-    const avgRenderTime =
-      this.renderTimes.length > 0
-        ? this.renderTimes.reduce((a, b) => a + b, 0) / this.renderTimes.length
-        : 0;
-
-    return {
-      fps: this.currentFps,
-      avgRenderTime: Math.round(avgRenderTime * 100) / 100,
-      droppedFrames: this.dropCount,
-      maxRenderTime: Math.max(...this.renderTimes, 0),
-    };
-  }
-}
-
-/**
- * Smart Frame Rate Controller
- * Synchronizes rendering with video source and manages performance
- */
-class FrameRateController {
-  private targetFps: number = 30;
-  private frameInterval: number;
-  private lastFrameTime = 0;
-  private animationId: number | null = null;
-  private isActive = false;
-  private monitor = new PerformanceMonitor();
-
-  constructor(private video: HTMLVideoElement, private onFrame: () => void) {
-    this.frameInterval = 1000 / this.targetFps;
-    this.detectVideoFps();
-  }
-
-  private async detectVideoFps(): Promise<void> {
-    try {
-      // Try to get actual video framerate
-      const track =
-        this.video.srcObject instanceof MediaStream
-          ? this.video.srcObject.getVideoTracks()[0]
-          : null;
-
-      if (track) {
-        const settings = track.getSettings();
-        if (settings.frameRate && settings.frameRate > 0) {
-          this.targetFps = Math.min(settings.frameRate, 60); // Cap at 60fps
-          this.frameInterval = 1000 / this.targetFps;
-          Logger.dev(`Detected video FPS: ${this.targetFps}`);
-        }
-      }
-    } catch (error) {
-      Logger.dev("FPS detection failed, using default:", error);
-    }
-  }
-
-  start(): void {
-    if (this.isActive) return;
-    this.isActive = true;
-    this.lastFrameTime = performance.now();
-    this.scheduleFrame();
-  }
-
-  stop(): void {
-    this.isActive = false;
-    if (this.animationId) {
-      cancelAnimationFrame(this.animationId);
-      this.animationId = null;
-    }
-  }
-
-  private scheduleFrame = (): void => {
-    if (!this.isActive) return;
-
-    const now = performance.now();
-    const elapsed = now - this.lastFrameTime;
-
-    if (elapsed >= this.frameInterval) {
-      const frameStart = this.monitor.startFrame();
-
-      try {
-        // Check if video is actually playing and has new frame
-        if (this.video.readyState >= 2 && !this.video.paused) {
-          this.onFrame();
-          this.monitor.endFrame(frameStart);
-        } else {
-          this.monitor.recordDrop();
-        }
-      } catch (error) {
-        Logger.dev("Frame render error:", error);
-        this.monitor.recordDrop();
-      }
-
-      this.lastFrameTime = now;
-    }
-
-    this.animationId = requestAnimationFrame(this.scheduleFrame);
-  };
-
-  getStats() {
-    return {
-      targetFps: this.targetFps,
-      ...this.monitor.getStats(),
-    };
-  }
-}
-
-/**
- * Configuration Cache for avoiding redundant calculations
- */
-class ConfigCache {
-  private cache = new Map<string, any>();
-
-  get<T>(key: string, factory: () => T): T {
-    if (!this.cache.has(key)) {
-      this.cache.set(key, factory());
-    }
-    return this.cache.get(key);
-  }
-
-  invalidate(key?: string): void {
-    if (key) {
-      this.cache.delete(key);
-    } else {
-      this.cache.clear();
-    }
-  }
-
-  size(): number {
-    return this.cache.size;
-  }
-}
-
-/**
- * Stream Processing Context
- * Manages the complete lifecycle of stream processing
- */
-class StreamProcessor {
-  private canvasResource: CanvasResource;
-  private frameController: FrameRateController;
-  private confettiManagers: Map<string, ConfettiOverlayManager> = new Map();
-  private mediaOverlayManagers: Map<string, MediaOverlayManager> = new Map();
-  private configCache = new ConfigCache();
-  private isDisposed = false;
-  private performanceMonitor = GlobalPerformanceMonitor.getInstance();
-  private streamId: string;
-  private errorCount = 0;
-  private maxErrors = 10;
-  private lastErrorTime = 0;
-  private errorThrottleMs = 1000;
-
-  constructor(
-    private video: HTMLVideoElement,
-    private crop: {
-      width: number;
-      height: number;
-      offsetX: number;
-      offsetY: number;
-    },
-    private filterConfig: StreamFilterConfig,
-    private cropConfig: StreamCropConfig
-  ) {
-    this.streamId = `stream_${Math.random().toString(36).substr(2, 9)}`;
-
-    try {
-      const canvasManager = CanvasManager.getInstance();
-      this.canvasResource = canvasManager.getCanvas(crop.width, crop.height);
-
-      this.setupCanvas();
-
-      this.frameController = new FrameRateController(video, this.renderFrame);
-
-      // Register with performance monitor
-      this.performanceMonitor.registerStream(this.streamId, {
-        fps: 0,
-        avgRenderTime: 0,
-        droppedFrames: 0,
-        maxRenderTime: 0,
-        activeStreams: 1,
-      });
-
-      // Register with global confetti manager
-      GlobalConfettiManager.getInstance().registerProcessor(this);
-
-      // Register with global media overlay manager
-      GlobalMediaOverlayManager.getInstance().registerProcessor(this);
-
-      // Register with global crop manager
-      GlobalCropManager.getInstance().registerProcessor(this);
-
-      // Register with global filter manager
-      GlobalFilterManager.getInstance().registerProcessor(this);
-
-      Logger.dev(`StreamProcessor initialized: ${this.streamId}`);
-    } catch (error) {
-      Logger.dev("StreamProcessor construction failed:", error);
-      throw error;
-    }
-  }
-
-  private setupCanvas(): void {
-    const { ctx } = this.canvasResource;
-
-    // Apply cached filter string using GlobalFilterManager
-    const filterManager = GlobalFilterManager.getInstance();
-    const filterKey = `filter_${JSON.stringify({
-      b: this.filterConfig.brightness,
-      s: this.filterConfig.saturation,
-      c: this.filterConfig.contrast,
-    })}`;
-
-    ctx.filter = this.configCache.get(filterKey, () =>
-      filterManager.generateFilterString(this.filterConfig)
-    );
-
-    // Apply mirror transformation using GlobalCropManager
-    const cropManager = GlobalCropManager.getInstance();
-    cropManager.applyMirrorTransform(
-      ctx,
-      this.crop.width,
-      this.cropConfig.mirror
-    );
-  }
-
-  private renderFrame = (): void => {
-    if (this.isDisposed) return;
-
-    const { ctx } = this.canvasResource;
-    const currentTime = performance.now();
-
-    try {
-      // Clear previous frame efficiently
-      ctx.clearRect(0, 0, this.crop.width, this.crop.height);
-
-      // Draw main video
-      ctx.drawImage(
-        this.video,
-        this.crop.offsetX,
-        this.crop.offsetY,
-        this.crop.width,
-        this.crop.height,
-        0,
-        0,
-        this.crop.width,
-        this.crop.height
-      );
-
-      // Optimized confetti rendering - batch all confetti operations
-      if (this.confettiManagers.size > 0) {
-        const finishedConfetti: string[] = [];
-
-        // Render all active confetti in one batch
-        ctx.save(); // Single save for all confetti
-
-        for (const [confettiId, confettiManager] of this.confettiManagers) {
-          if (confettiManager.shouldRender(currentTime)) {
-            confettiManager.render(ctx, currentTime);
-          } else {
-            finishedConfetti.push(confettiId);
-          }
-        }
-
-        ctx.restore(); // Single restore for all confetti
-
-        // Clean up finished confetti
-        for (const confettiId of finishedConfetti) {
-          const confettiManager = this.confettiManagers.get(confettiId);
-          if (confettiManager) {
-            confettiManager.cleanup();
-            this.confettiManagers.delete(confettiId);
-          }
-        }
-      }
-
-      // Media overlay rendering
-      if (this.mediaOverlayManagers.size > 0) {
-        const finishedOverlays: string[] = [];
-
-        for (const [overlayId, overlayManager] of this.mediaOverlayManagers) {
-          if (overlayManager.shouldRender(currentTime)) {
-            overlayManager.render(ctx, currentTime);
-          } else {
-            finishedOverlays.push(overlayId);
-          }
-        }
-
-        // Clean up finished overlays
-        for (const overlayId of finishedOverlays) {
-          const overlayManager = this.mediaOverlayManagers.get(overlayId);
-          if (overlayManager) {
-            overlayManager.cleanup();
-            this.mediaOverlayManagers.delete(overlayId);
-          }
-        }
-      }
-
-      // Reset error count on successful render
-      this.errorCount = 0;
-    } catch (error) {
-      this.handleRenderError(error);
-    }
-  };
-
-  private handleRenderError(error: unknown): void {
-    const now = performance.now();
-
-    // Throttle error logging to prevent spam
-    if (now - this.lastErrorTime > this.errorThrottleMs) {
-      Logger.dev(`Render error in ${this.streamId}:`, error);
-      this.lastErrorTime = now;
-    }
-
-    this.errorCount++;
-
-    // If too many errors, dispose the processor to prevent resource waste
-    if (this.errorCount >= this.maxErrors) {
-      Logger.dev(`Too many errors in ${this.streamId}, disposing processor`);
-      this.dispose();
-      return;
-    }
-
-    // Try to recover from certain types of errors
-    if (error instanceof Error) {
-      const errorMsg = error.message.toLowerCase();
-
-      // Handle canvas context loss
-      if (errorMsg.includes("context") || errorMsg.includes("canvas")) {
-        this.attemptCanvasRecovery();
-      }
-
-      // Handle video element issues
-      if (errorMsg.includes("video") || errorMsg.includes("media")) {
-        this.attemptVideoRecovery();
-      }
-    }
-  }
-
-  private attemptCanvasRecovery(): void {
-    try {
-      const canvasManager = CanvasManager.getInstance();
-      const oldResource = this.canvasResource;
-
-      // Try to get a new canvas resource
-      this.canvasResource = canvasManager.getCanvas(
-        this.crop.width,
-        this.crop.height
-      );
-      this.setupCanvas();
-
-      // Release the old resource
-      canvasManager.releaseCanvas(oldResource);
-
-      Logger.dev(`Canvas recovery successful for ${this.streamId}`);
-    } catch (recoveryError) {
-      Logger.dev(`Canvas recovery failed for ${this.streamId}:`, recoveryError);
-    }
-  }
-
-  private attemptVideoRecovery(): void {
-    try {
-      // Check if video element is still valid
-      if (this.video.readyState === 0) {
-        // Video element lost its source, try to reload
-        this.video.load();
-      }
-
-      Logger.dev(`Video recovery attempted for ${this.streamId}`);
-    } catch (recoveryError) {
-      Logger.dev(`Video recovery failed for ${this.streamId}:`, recoveryError);
-    }
-  }
-
-  startConfetti(
-    confettiId: string,
-    config: {
-      confettiType: string;
-      colors: string[];
-      intensity: number;
-      duration: number;
-    }
-  ): void {
-    if (this.isDisposed) return;
-
-    const confettiManager = new ConfettiOverlayManager(config, {
-      width: this.crop.width,
-      height: this.crop.height,
-    });
-
-    confettiManager.start();
-    this.confettiManagers.set(confettiId, confettiManager);
-
-    Logger.dev(`Confetti started in ${this.streamId}: ${confettiId}`);
-  }
-
-  stopConfetti(confettiId: string): void {
-    const confettiManager = this.confettiManagers.get(confettiId);
-    if (confettiManager) {
-      confettiManager.stop();
-      confettiManager.cleanup();
-      this.confettiManagers.delete(confettiId);
-      Logger.dev(`Confetti stopped in ${this.streamId}: ${confettiId}`);
-    }
-  }
-
-  clearAllConfetti(): void {
-    this.confettiManagers.forEach((confettiManager) => {
-      confettiManager.stop();
-      confettiManager.cleanup();
-    });
-    this.confettiManagers.clear();
-    Logger.dev(`All confetti cleared in ${this.streamId}`);
-  }
-
-  startMediaOverlay(
-    overlayId: string,
-    config: {
-      mediaUrl: string;
-      mediaType: "video" | "image";
-      position: { x: number; y: number };
-      scale: number;
-      duration: number;
-      opacity: number;
-      delay: number;
-    }
-  ): void {
-    if (this.isDisposed) return;
-
-    const mediaOverlayManager = new MediaOverlayManager(config, {
-      width: this.crop.width,
-      height: this.crop.height,
-    });
-
-    mediaOverlayManager.start();
-    this.mediaOverlayManagers.set(overlayId, mediaOverlayManager);
-
-    Logger.dev(`Media overlay started in ${this.streamId}: ${overlayId}`);
-  }
-
-  stopMediaOverlay(overlayId: string): void {
-    const mediaOverlayManager = this.mediaOverlayManagers.get(overlayId);
-    if (mediaOverlayManager) {
-      mediaOverlayManager.stop();
-      mediaOverlayManager.cleanup();
-      this.mediaOverlayManagers.delete(overlayId);
-      Logger.dev(`Media overlay stopped in ${this.streamId}: ${overlayId}`);
-    }
-  }
-
-  clearAllMediaOverlays(): void {
-    this.mediaOverlayManagers.forEach((mediaOverlayManager) => {
-      mediaOverlayManager.stop();
-      mediaOverlayManager.cleanup();
-    });
-    this.mediaOverlayManagers.clear();
-    Logger.dev(`All media overlays cleared in ${this.streamId}`);
-  }
-
-  updateCropSettings(config: {
-    aspectRatio?: number;
-    zoom?: number;
-    align?: "left" | "center" | "right";
-    mirror?: boolean;
-  }): void {
-    if (this.isDisposed) return;
-
-    // Update config with new crop settings
-    this.filterConfig = { ...this.filterConfig, ...config };
-
-    // Recalculate crop dimensions
-    const cropManager = GlobalCropManager.getInstance();
-    const newCrop = cropManager.calculateCrop(
-      {
-        width: this.canvasResource.canvas.width,
-        height: this.canvasResource.canvas.height,
-      },
-      config.aspectRatio
-    );
-    const newZoom = cropManager.calculateZoomedSize(newCrop, config.zoom);
-    const newOffset = cropManager.calculateOffset(
-      {
-        width: this.canvasResource.canvas.width,
-        height: this.canvasResource.canvas.height,
-      },
-      newZoom,
-      config.align
-    );
-
-    // Update crop settings
-    this.crop = {
-      ...newZoom,
-      offsetX: newOffset.x,
-      offsetY: newOffset.y,
-    };
-
-    // Clear configuration cache to force recalculation
-    this.configCache.invalidate();
-
-    // If mirror setting changed, we need to reset the canvas transform
-    if (config.mirror !== undefined) {
-      const { ctx } = this.canvasResource;
-      ctx.resetTransform(); // Reset all transforms
-      this.setupCanvas(); // Reapply all canvas settings including mirror
-    }
-
-    Logger.dev(`Crop settings updated for ${this.streamId}:`, this.crop);
-  }
-
-  updateFilterSettings(config: StreamFilterConfig): void {
-    if (this.isDisposed) return;
-
-    // Update filter config with new settings
-    this.filterConfig = { ...this.filterConfig, ...config };
-
-    // Clear configuration cache to force recalculation
-    this.configCache.invalidate();
-
-    // Reapply canvas filters
-    const { ctx } = this.canvasResource;
-    const filterManager = GlobalFilterManager.getInstance();
-    ctx.filter = filterManager.generateFilterString(this.filterConfig);
-
-    Logger.dev(
-      `Filter settings updated for ${this.streamId}:`,
-      this.filterConfig
-    );
-  }
-
-  start(): HTMLCanvasElement {
-    this.frameController.start();
-    return this.canvasResource.canvas;
-  }
-
-  getStats() {
-    const frameStats = this.frameController.getStats();
-    const stats = {
-      ...frameStats,
-      cacheSize: this.configCache.size(),
-      canvasStats: CanvasManager.getInstance().getStats(),
-      errorCount: this.errorCount,
-      streamId: this.streamId,
-    };
-
-    // Update global performance monitor
-    this.performanceMonitor.updateMetrics(this.streamId, {
-      fps: frameStats.fps,
-      avgRenderTime: frameStats.avgRenderTime,
-      droppedFrames: frameStats.droppedFrames,
-      maxRenderTime: frameStats.maxRenderTime,
-      activeStreams: 1,
-    });
-
-    return stats;
-  }
-
-  dispose(): void {
-    if (this.isDisposed) return;
-    this.isDisposed = true;
-
-    try {
-      // Stop frame controller
-      this.frameController.stop();
-
-      // Cleanup all confetti overlays
-      this.confettiManagers.forEach((confettiManager) => {
-        confettiManager.cleanup();
-      });
-      this.confettiManagers.clear();
-
-      // Cleanup all media overlays
-      this.mediaOverlayManagers.forEach((mediaOverlayManager) => {
-        mediaOverlayManager.cleanup();
-      });
-      this.mediaOverlayManagers.clear();
-
-      // Unregister from global confetti manager
-      GlobalConfettiManager.getInstance().unregisterProcessor(this);
-
-      // Unregister from global media overlay manager
-      GlobalMediaOverlayManager.getInstance().unregisterProcessor(this);
-
-      // Unregister from global crop manager
-      GlobalCropManager.getInstance().unregisterProcessor(this);
-
-      // Unregister from global filter manager
-      GlobalFilterManager.getInstance().unregisterProcessor(this);
-
-      // Clear configuration cache
-      this.configCache.invalidate();
-
-      // Release canvas resource
-      const canvasManager = CanvasManager.getInstance();
-      canvasManager.releaseCanvas(this.canvasResource);
-
-      // Unregister from performance monitor
-      this.performanceMonitor.removeStream(this.streamId);
-
-      Logger.dev(`StreamProcessor disposed: ${this.streamId}`);
-    } catch (error) {
-      Logger.dev(
-        `Error during StreamProcessor disposal: ${this.streamId}`,
-        error
-      );
-    }
-  }
-}
-
-function applyCanvasProcessing(settings: {
-  video: HTMLVideoElement;
-  crop: { width: number; height: number; offsetX: number; offsetY: number };
-  filterConfig: StreamFilterConfig;
-  cropConfig: StreamCropConfig;
-}): { canvas: HTMLCanvasElement; processor: StreamProcessor } {
-  const { video, crop, filterConfig, cropConfig } = settings;
-
-  const processor = new StreamProcessor(video, crop, filterConfig, cropConfig);
-  const canvas = processor.start();
-
-  return { canvas, processor };
-}
-
-function setupVideoElement(track: MediaStreamTrack): HTMLVideoElement {
-  const video = document.createElement("video");
-  video.srcObject = new MediaStream([track]);
-  video.muted = true;
-  video.playsInline = true;
-  video.autoplay = true;
-  video.play().catch((err) => {
-    Logger.dev("Video play error:", err);
-  });
-  return video;
-}
-
-/**
- * Global confetti manager for handling confetti triggers across all streams
- */
-class GlobalConfettiManager {
-  private static instance: GlobalConfettiManager;
-  private processors: Set<StreamProcessor> = new Set();
-
-  static getInstance(): GlobalConfettiManager {
-    if (!GlobalConfettiManager.instance) {
-      GlobalConfettiManager.instance = new GlobalConfettiManager();
-    }
-    return GlobalConfettiManager.instance;
-  }
-
-  registerProcessor(processor: StreamProcessor): void {
-    this.processors.add(processor);
-  }
-
-  unregisterProcessor(processor: StreamProcessor): void {
-    this.processors.delete(processor);
-  }
-
-  triggerConfetti(config: {
-    confettiType: string;
-    colors: string[];
-    intensity: number;
-    duration: number;
-  }): void {
-    const confettiId = `confetti_${Date.now()}`;
-    Logger.dev("Triggering confetti for all active streams:", config);
-
-    // Clear all existing confetti first to prevent stacking
-    this.processors.forEach((processor) => {
-      processor.clearAllConfetti();
-    });
-
-    // Start new confetti
-    this.processors.forEach((processor) => {
-      processor.startConfetti(confettiId, config);
-    });
-
-    // Auto-cleanup after duration
-    setTimeout(() => {
-      this.processors.forEach((processor) => {
-        processor.stopConfetti(confettiId);
-      });
-    }, config.duration * 1000 + 1000);
-  }
-}
-
-/**
- * Global media overlay manager for handling media overlay triggers across all streams
- */
-class GlobalMediaOverlayManager {
-  private static instance: GlobalMediaOverlayManager;
-  private processors: Set<StreamProcessor> = new Set();
-
-  static getInstance(): GlobalMediaOverlayManager {
-    if (!GlobalMediaOverlayManager.instance) {
-      GlobalMediaOverlayManager.instance = new GlobalMediaOverlayManager();
-    }
-    return GlobalMediaOverlayManager.instance;
-  }
-
-  registerProcessor(processor: StreamProcessor): void {
-    this.processors.add(processor);
-  }
-
-  unregisterProcessor(processor: StreamProcessor): void {
-    this.processors.delete(processor);
-  }
-
-  triggerMediaOverlay(config: {
-    mediaUrl: string;
-    mediaType: "video" | "image";
-    position: { x: number; y: number };
-    scale: number;
-    duration: number;
-    opacity: number;
-    delay: number;
-  }): void {
-    const overlayId = `media_overlay_${Date.now()}`;
-    Logger.dev("Triggering media overlay for all active streams:", config);
-
-    // Clear all existing media overlays first to prevent stacking
-    this.processors.forEach((processor) => {
-      processor.clearAllMediaOverlays();
-    });
-
-    // Start new media overlay
-    this.processors.forEach((processor) => {
-      processor.startMediaOverlay(overlayId, config);
-    });
-
-    // Auto-cleanup after duration + delay
-    setTimeout(() => {
-      this.processors.forEach((processor) => {
-        processor.stopMediaOverlay(overlayId);
-      });
-    }, (config.duration + config.delay) * 1000 + 1000);
-  }
-}
-
-/**
- * Global crop manager for handling crop settings across all streams
- */
-class GlobalCropManager {
-  private static instance: GlobalCropManager;
-  private processors: Set<StreamProcessor> = new Set();
-  private currentCropConfig: {
-    aspectRatio?: number;
-    zoom?: number;
-    align?: "left" | "center" | "right";
-    mirror?: boolean;
-  } = {};
-
-  static getInstance(): GlobalCropManager {
-    if (!GlobalCropManager.instance) {
-      GlobalCropManager.instance = new GlobalCropManager();
-    }
-    return GlobalCropManager.instance;
-  }
-
-  registerProcessor(processor: StreamProcessor): void {
-    this.processors.add(processor);
-  }
-
-  unregisterProcessor(processor: StreamProcessor): void {
-    this.processors.delete(processor);
-  }
-
-  calculateCrop(size: StreamPatcherSize, aspectRatio?: number) {
-    const { width, height } = size;
-
-    if (typeof aspectRatio !== "number") {
-      return {
-        width: width,
-        height: height,
-        offsetX: 0,
-        offsetY: 0,
-      };
-    }
-
-    const originalAspect = width / height;
-    let cropWidth = width;
-    let cropHeight = height;
-
-    if (originalAspect > aspectRatio) {
-      cropWidth = Math.floor(height * aspectRatio);
-    } else {
-      cropHeight = Math.floor(width / aspectRatio);
-    }
-
-    Logger.dev(`Crop: ${cropWidth}x${cropHeight}`);
-    return { width: cropWidth, height: cropHeight };
-  }
-
-  calculateZoomedSize(size: StreamPatcherSize, zoom?: number) {
-    if (typeof zoom !== "number" || zoom <= 1) {
-      return { ...size };
-    }
-
-    const zoomWidth = Math.floor(size.width / zoom);
-    const zoomHeight = Math.floor(size.height / zoom);
-
-    Logger.dev(`zoom: ${zoomWidth}x${zoomHeight}`);
-
-    return {
-      width: zoomWidth,
-      height: zoomHeight,
-    };
-  }
-
-  calculateOffset(
-    original: StreamPatcherSize,
-    modified: StreamPatcherSize,
-    align: StreamCropConfig["align"] = "center"
-  ) {
-    let x = 0;
-
-    if (align === "left") {
-      x = 0;
-    } else if (align === "right") {
-      x = original.width - modified.width;
-    } else {
-      x = Math.floor((original.width - modified.width) / 2);
-    }
-
-    const y = Math.floor((original.height - modified.height) / 2);
-
-    return { x, y };
-  }
-
-  applyMirrorTransform(
-    ctx: CanvasRenderingContext2D,
-    cropWidth: number,
-    mirror?: boolean
-  ): void {
-    if (mirror) {
-      ctx.scale(-1, 1);
-      ctx.translate(-cropWidth, 0);
-    }
-  }
-
-  initProcessor(
-    size: StreamPatcherSize,
-    config: StreamCropConfig = {}
-  ): {
-    crop: { width: number; height: number; offsetX: number; offsetY: number };
-    cropConfig: StreamCropConfig;
-  } {
-    // Merge provided config with current state, giving precedence to provided values
-    const mergedConfig = {
-      aspectRatio: config.aspectRatio ?? this.currentCropConfig.aspectRatio,
-      zoom: config.zoom ?? this.currentCropConfig.zoom,
-      align: config.align ?? this.currentCropConfig.align,
-      mirror: config.mirror ?? this.currentCropConfig.mirror,
-    };
-
-    const crop = this.calculateCrop(size, mergedConfig.aspectRatio);
-    const zoom = this.calculateZoomedSize(crop, mergedConfig.zoom);
-    const offset = this.calculateOffset(size, zoom, mergedConfig.align);
-
-    // Update current config with merged values
-    this.currentCropConfig = { ...mergedConfig };
-
-    return {
-      crop: {
-        ...zoom,
-        offsetX: offset.x,
-        offsetY: offset.y,
-      },
-      cropConfig: mergedConfig,
-    };
-  }
-
-  applyCrop(config: StreamCropConfig): void {
-    this.currentCropConfig = { ...this.currentCropConfig, ...config };
-    Logger.dev("Applying crop settings to all active streams:", config);
-
-    // Apply crop settings to all processors
-    this.processors.forEach((processor) => {
-      processor.updateCropSettings(config);
-    });
-  }
-
-  getCurrentCropConfig() {
-    return { ...this.currentCropConfig };
-  }
-}
-
-/**
- * Global filter manager for handling filter settings across all streams
- */
-class GlobalFilterManager {
-  private static instance: GlobalFilterManager;
-  private processors: Set<StreamProcessor> = new Set();
-  private currentFilterConfig: StreamFilterConfig = {};
-
-  static getInstance(): GlobalFilterManager {
-    if (!GlobalFilterManager.instance) {
-      GlobalFilterManager.instance = new GlobalFilterManager();
-    }
-    return GlobalFilterManager.instance;
-  }
-
-  registerProcessor(processor: StreamProcessor): void {
-    this.processors.add(processor);
-  }
-
-  unregisterProcessor(processor: StreamProcessor): void {
-    this.processors.delete(processor);
-  }
-
-  generateFilterString(config: StreamFilterConfig): string {
-    const brightness = normalizeFilterValue(config.brightness);
-    const saturation = normalizeFilterValue(config.saturation);
-    const contrast = normalizeFilterValue(config.contrast);
-
-    return `brightness(${brightness}) saturate(${saturation}) contrast(${contrast})`;
-  }
-
-  initProcessor(config: StreamFilterConfig = {}): {
-    filterConfig: StreamFilterConfig;
-    filterString: string;
-  } {
-    // Merge provided config with current state, giving precedence to provided values
-    const mergedConfig = {
-      brightness: config.brightness ?? this.currentFilterConfig.brightness,
-      saturation: config.saturation ?? this.currentFilterConfig.saturation,
-      contrast: config.contrast ?? this.currentFilterConfig.contrast,
-    };
-
-    const filterString = this.generateFilterString(mergedConfig);
-
-    // Update current config with merged values
-    this.currentFilterConfig = { ...mergedConfig };
-
-    return {
-      filterConfig: mergedConfig,
-      filterString,
-    };
-  }
-
-  applyFilters(config: StreamFilterConfig): void {
-    this.currentFilterConfig = { ...this.currentFilterConfig, ...config };
-    Logger.dev("Applying filter settings to all active streams:", config);
-
-    // Apply filter settings to all processors
-    this.processors.forEach((processor) => {
-      processor.updateFilterSettings(config);
-    });
-  }
-
-  getCurrentFilterConfig() {
-    return { ...this.currentFilterConfig };
-  }
-}
-
-/**
- * Stream processors registry for cleanup management
- */
-const activeProcessors = new WeakMap<MediaStream, StreamProcessor>();
-
-export function streamPatcher(
-  settings: StreamPatcherSettings,
-  stopOriginalStream = false
-): MediaStream {
-  const { cropConfig, filterConfig, size, stream } = settings;
-
-  try {
-    const videoTrack = stream.getVideoTracks()[0];
-    if (!videoTrack) throw new Error("No video track found in stream.");
-
-    const video = setupVideoElement(videoTrack);
-
-    // Use GlobalCropManager for initialization
-    const cropManager = GlobalCropManager.getInstance();
-    const { crop } = cropManager.initProcessor(size, cropConfig);
-
-    // Use GlobalFilterManager for initialization
-    const filterManager = GlobalFilterManager.getInstance();
-    filterManager.initProcessor(filterConfig);
-
-    const { canvas, processor } = applyCanvasProcessing({
-      video,
-      crop,
-      filterConfig,
-      cropConfig,
-    });
-
-    // Create output stream with explicit framerate for better performance
-    const targetFps = Math.min(30, (filterConfig as any).maxFps || 30);
-    const outputStream = canvas.captureStream(targetFps);
-
-    // Store processor reference for cleanup
-    activeProcessors.set(outputStream, processor);
-
-    // Enhanced cleanup management
-    if (stopOriginalStream) {
-      outputStream.getTracks().forEach((track: MediaStreamTrack) => {
-        const originalStop = track.stop.bind(track);
-        track.stop = () => {
-          try {
-            // Clean up processor resources
-            const processor = activeProcessors.get(outputStream);
-            if (processor) {
-              processor.dispose();
-              activeProcessors.delete(outputStream);
-            }
-
-            // Stop original stream
-            stream.getTracks().forEach((t) => t.stop());
-            Logger.dev("Stream and processor cleaned up successfully");
-          } catch (err) {
-            Logger.dev("Error during cleanup:", err);
-          } finally {
-            originalStop();
-          }
-        };
-      });
-    } else {
-      // Set up cleanup for when output stream is stopped
-      outputStream.getTracks().forEach((track: MediaStreamTrack) => {
-        const originalStop = track.stop.bind(track);
-        track.stop = () => {
-          const processor = activeProcessors.get(outputStream);
-          if (processor) {
-            processor.dispose();
-            activeProcessors.delete(outputStream);
-          }
-          originalStop();
-        };
-      });
-    }
-
-    // Log performance stats in development
-    if (
-      typeof process !== "undefined" &&
-      process.env?.NODE_ENV === "development"
-    ) {
-      setTimeout(() => {
-        const stats = processor.getStats();
-        Logger.dev("Stream processor stats:", stats);
-      }, 5000);
-    }
-
-    return outputStream;
-  } catch (err) {
-    Logger.dev("streamPatcher error:", err);
-    return stream; // Fallback to original stream if something fails
-  }
-}
-
-/**
- * Utility function to cleanup all active processors
- * Useful for extension cleanup or page navigation
- */
-export function cleanupAllProcessors(): void {
-  // Note: WeakMap doesn't have iteration methods
-  // Processors will be cleaned up automatically when streams are garbage collected
-  const canvasManager = CanvasManager.getInstance();
-  canvasManager.dispose();
-  Logger.dev("All stream processors cleanup initiated");
-}
-
-/**
- * Function to trigger confetti across all active video streams
- */
-export function triggerGlobalConfetti(config: {
-  confettiType: string;
-  colors: string[];
-  intensity: number;
-  duration: number;
-}): void {
-  GlobalConfettiManager.getInstance().triggerConfetti(config);
-}
-
-/**
- * Function to trigger media overlay across all active video streams
- */
-export function triggerGlobalMediaOverlay(config: {
-  mediaUrl: string;
-  mediaType: "video" | "image";
-  position: { x: number; y: number };
-  scale: number;
-  duration: number;
-  opacity: number;
-  delay: number;
-}): void {
-  GlobalMediaOverlayManager.getInstance().triggerMediaOverlay(config);
-}
-
-/**
- * Function to apply crop settings across all active video streams
- */
-export function applyGlobalCrop(config: StreamCropConfig): void {
-  GlobalCropManager.getInstance().applyCrop(config);
-}
-
-/**
- * Function to apply filter settings across all active video streams
- */
-export function applyGlobalFilters(config: StreamFilterConfig): void {
-  GlobalFilterManager.getInstance().applyFilters(config);
-}
