+++ src/utils/stream-patcher/confetti-overlay.ts
@@ -1,4 +1,4 @@
-import { Logger } from "./log";
+import { Logger } from "@/utils/log";
 
 interface ConfettiParticle {
   x: number;
