+++ src/utils/stream-patcher/media-overlay.ts
@@ -1,4 +1,4 @@
-import { Logger } from "./log";
+import { Logger } from "@/utils/log";
 
 interface MediaOverlayConfig {
   mediaUrl: string;
@@ -69,8 +69,11 @@ export class MediaOverlayManager {
     this.isDelaying = this.config.delay > 0;
     this.delayStartTime = performance.now();
     this.startTime = performance.now();
-    
-    if (this.config.mediaType === "video" && this.mediaElement instanceof HTMLVideoElement) {
+
+    if (
+      this.config.mediaType === "video" &&
+      this.mediaElement instanceof HTMLVideoElement
+    ) {
       this.mediaElement.play().catch((error) => {
         Logger.dev("Failed to play video:", error);
       });
@@ -109,10 +112,16 @@ export class MediaOverlayManager {
     }
 
     // Only render if media is loaded
-    if (this.config.mediaType === "image" && !(this.mediaElement as HTMLImageElement).complete) {
+    if (
+      this.config.mediaType === "image" &&
+      !(this.mediaElement as HTMLImageElement).complete
+    ) {
       return;
     }
-    if (this.config.mediaType === "video" && (this.mediaElement as HTMLVideoElement).readyState < 2) {
+    if (
+      this.config.mediaType === "video" &&
+      (this.mediaElement as HTMLVideoElement).readyState < 2
+    ) {
       return;
     }
 
@@ -122,12 +131,14 @@ export class MediaOverlayManager {
     ctx.globalAlpha = this.config.opacity / 100;
 
     // Get original media dimensions
-    const mediaWidth = this.config.mediaType === "image" 
-      ? (this.mediaElement as HTMLImageElement).naturalWidth 
-      : (this.mediaElement as HTMLVideoElement).videoWidth;
-    const mediaHeight = this.config.mediaType === "image" 
-      ? (this.mediaElement as HTMLImageElement).naturalHeight 
-      : (this.mediaElement as HTMLVideoElement).videoHeight;
+    const mediaWidth =
+      this.config.mediaType === "image"
+        ? (this.mediaElement as HTMLImageElement).naturalWidth
+        : (this.mediaElement as HTMLVideoElement).videoWidth;
+    const mediaHeight =
+      this.config.mediaType === "image"
+        ? (this.mediaElement as HTMLImageElement).naturalHeight
+        : (this.mediaElement as HTMLVideoElement).videoHeight;
 
     if (mediaWidth === 0 || mediaHeight === 0) {
       ctx.restore();
@@ -136,13 +147,14 @@ export class MediaOverlayManager {
 
     // Calculate size based on stream dimensions, similar to position-picker
     // Use a base size relative to the smaller dimension of the canvas (like position-picker)
-    const baseSize = Math.min(this.canvasSize.width, this.canvasSize.height) * 0.2; // 20% of smaller dimension
-    
+    const baseSize =
+      Math.min(this.canvasSize.width, this.canvasSize.height) * 0.2; // 20% of smaller dimension
+
     // Maintain aspect ratio while scaling
     const aspectRatio = mediaWidth / mediaHeight;
     let scaledWidth: number;
     let scaledHeight: number;
-    
+
     if (aspectRatio >= 1) {
       // Landscape or square
       scaledWidth = baseSize * this.config.scale * aspectRatio;
@@ -150,12 +162,15 @@ export class MediaOverlayManager {
     } else {
       // Portrait
       scaledWidth = baseSize * this.config.scale;
-      scaledHeight = baseSize * this.config.scale / aspectRatio;
+      scaledHeight = (baseSize * this.config.scale) / aspectRatio;
     }
 
     // Convert percentage position to canvas coordinates (centered)
-    const x = (this.config.position.x / 100) * this.canvasSize.width - scaledWidth / 2;
-    const y = (this.config.position.y / 100) * this.canvasSize.height - scaledHeight / 2;
+    const x =
+      (this.config.position.x / 100) * this.canvasSize.width - scaledWidth / 2;
+    const y =
+      (this.config.position.y / 100) * this.canvasSize.height -
+      scaledHeight / 2;
 
     // Draw the media
     ctx.drawImage(this.mediaElement, x, y, scaledWidth, scaledHeight);
@@ -165,7 +180,10 @@ export class MediaOverlayManager {
 
   stop(): void {
     this.isActive = false;
-    if (this.config.mediaType === "video" && this.mediaElement instanceof HTMLVideoElement) {
+    if (
+      this.config.mediaType === "video" &&
+      this.mediaElement instanceof HTMLVideoElement
+    ) {
       this.mediaElement.pause();
     }
     Logger.dev(`Media overlay stopped: ${this.config.mediaUrl}`);
@@ -174,7 +192,10 @@ export class MediaOverlayManager {
   cleanup(): void {
     this.stop();
     if (this.mediaElement) {
-      if (this.config.mediaType === "video" && this.mediaElement instanceof HTMLVideoElement) {
+      if (
+        this.config.mediaType === "video" &&
+        this.mediaElement instanceof HTMLVideoElement
+      ) {
         this.mediaElement.pause();
         this.mediaElement.src = "";
       }
