+++ src/utils/stream-patcher/performance-monitor.ts
