+++ src/utils/stream-patcher/types.ts
